implement carts index function
