+++ app/controllers/products_controller.rb
@@ -14,21 +14,4 @@ class ProductsController < ApplicationController
     flash[:notice] = "成功加入购物车"
   end
 
-  helper_method :current_cart
-
-  def current_cart
-    @current_cart ||= find_cart
-  end
-
-  private
-
-  def find_cart
-    cart = Cart.find_by(id: session[:cart_id])
-    if cart.blank?
-      cart = Cart.create
-    end
-
-    session[:cart_id] = cart.id
-    return cart
-  end
 end
