+++ config/routes.rb
@@ -11,4 +11,6 @@ Rails.application.routes.draw do
       post :add_to_cart
     end
   end
+
+  resources :carts
 end
